frac

+++ main.cu
@@ -208,7 +208,7 @@ int main(int argc, char ** argv)
         if (strcmp(argv[i], "--max_iters") == 0)
             max_iters = atoi(argv[i+1]);  
         if (strcmp(argv[i], "--frac") == 0)
-            fraction = atoi(argv[i+1]);      
+            fraction = atof(argv[i+1]);      
     }
 
     std::cout << "frac" << "\t" << "Iter" << "\t" << "Trav" << "\t" << "isColored" << "\t"<< "Total" << "\t" << "isRight" << "\t"<< "Colors" << endl;
